+++ javascript/login.js
@@ -1,14 +1,14 @@
-const form = document.querySelector('.signup form');
-const signupBtn = form.querySelector('.buttom input');
+const form = document.querySelector('.login form');
+const loginBtn = form.querySelector('.buttom input');
 const errorTxt = form.querySelector('.error-txt');
 
 form.onsubmit = (e) => {
     e.preventDefault();
 }
 
-signupBtn.onclick = () => {
+loginBtn.onclick = () => {
     let xhr = new XMLHttpRequest();
-    xhr.open("POST", "php/register.php");
+    xhr.open("POST", "php/login.php");
     xhr.onload = () => {
         if (xhr.readyState === XMLHttpRequest.DONE) {
             if (xhr.status === 200) {
@@ -29,11 +29,6 @@ signupBtn.onclick = () => {
     }
 
     let formData = new FormData(form);
-    let active_img = document.querySelector(".signup form .image li img.active");
-    if (active_img) {
-        let image_src = active_img.src.split('/').pop();
-        formData.append('image', image_src);
-    }
 
     xhr.send(formData);
 }