+++ login.php
@@ -13,21 +13,22 @@
                 <div class="error-txt">This is an error message!!</div>
                 <div class="field input">
                     <label>Email Address</label>
-                    <input type="text" placeholder="Enter your email" required>
+                    <input type="text" placeholder="Enter your email" name="email" required>
                 </div>
                 <div class="field input">
                     <label>Password</label>
-                    <input type="password" placeholder="Enter your password" required>
+                    <input type="password" placeholder="Enter your password" name="password" required>
                     <i class="fas fa-eye"></i>
                 </div>
                 <div class="field buttom">
                     <input type="submit" value="Continue to Chat">
                 </div>
             </form>
-            <div class="link">Not yet signed up? <a href="index.html">Signup now</a></div>
+            <div class="link">Not yet signed up? <a href="index.php">Signup now</a></div>
         </section>
     </div>
 
     <script src="javascript/pass-show-hide.js"></script>
+    <script src="javascript/login.js"></script>
 </body>
 </html>